Parser

+++ src/compiler/types/text_range.rs
@@ -2,12 +2,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Eq, PartialEq, Debug, Copy, Clone, Deserialize, Serialize)]
 pub struct TextRange {
-    pos: Option<u32>,
-    end: Option<u32>,
+    pos: Option<usize>,
+    end: Option<usize>,
 }
 
 impl TextRange {
-    pub fn from_pos_and_end(pos: Option<u32>, end: Option<u32>) -> TextRange {
+    pub fn from_pos_and_end(pos: Option<usize>, end: Option<usize>) -> TextRange {
         if let (Some(pos), Some(end)) = (pos, end) {
             assert!(end >= pos);
         }
